make transaction status return an either
